Bump MIN_UBOOT_FIRM_VER from v2.10.4 to v2.11.5
The previous versions of firmware including v2.10.4 had a bug related to
power management where the board would not reboot properly. Just when
the board would turn off, the voltage would drop from 10-ish volts to
5v. This would happen randomly and prevent the board from powering back
up.

Firmware provided in release v2.11.5 fixes this. This was included in the
changelog and I have verified it using all the compatible cables that I
have. More testing is needed to make sure that *this* firmware version
fixes it or not. But making this change because more community members
and/or testers use this as a minimum version and raise issues in regards
of this version.