docs/riscv64/visionfive2/visionfive2.md: finalize release version
We have set a minimum firmware version for the VisionFive 2 SBC to
v2.11.5, so now it is time to replace all the placeholders and
"finalize" the links.

The link for the file "sdcard.img" is deliberately set to the release
version v2.8.0 primarily because that is what I used to update both of
my board's firmware. Meaning, it is tested and works as intended.

If someone else from the AltArch SIG uses the sdcard.img file from a
newer release and confirms that they are able to boot it on the older
firmware and update the firmware, by all means, update the version
string on line 72.